feat: move config code to internal
